Smol improvements
